Make version calculation faster and rely only on action names

+++ Source/DynamicOpenVR/Manifest/ActionManifest.cs
@@ -22,7 +22,7 @@ namespace DynamicOpenVR.Manifest
     internal class ActionManifest
     {
         [JsonProperty(PropertyName = "version")]
-        internal ulong version { get; set; } = 1;
+        internal int version { get; set; } = 1;
 
         [JsonProperty(PropertyName = "actions")]
         internal List<ManifestAction> actions { get; set; } = new List<ManifestAction>();
